fix(tools): 修复代码重复问题 (#54) (#58)
- 将 _validate_mtime 和 _generate_diff 方法从 EditTool 和 WriteTool 移至基类 BaseTool，消除重复代码
- 在 BaseTool 中添加 handle_tool_exceptions 装饰器，统一处理工具方法的异常
- 使用装饰器替换所有工具类中的四段式异常处理代码：
  - edit_tool.py, write_tool.py, read_tool.py, read_lines_tool.py
  - ls_tool.py, glob_tool.py, stat_tool.py, regex_search_tool.py
  - exact_search_tool.py, symbol_ref_tool.py
- 移除工具类中无用的导入和重复代码
