+++ src/workspace/tools/glob_tool.py
@@ -12,6 +12,7 @@ class GlobTool(BaseTool):
         self.func = self.glob
         self.params = BaseTool.extract_params(self.glob)
 
+    @BaseTool.handle_tool_exceptions
     def glob(self, pattern: str, folder_path: str = ".", max_ret: int = 1000) -> list[str]:
         """
         在工作区内按通配符模式匹配并列出所有路径,带[Folder]或[File]的类型标记. 失败时返回错误信息
@@ -26,20 +27,11 @@ class GlobTool(BaseTool):
         -------
         检索到的文件或文件夹的相对路径
         """
-        try:
-            root_path = self.workspace.path_validator.validate(folder_path)
-            if not root_path.is_dir():
-                return ToolErrorResponse(self.__class__.__name__, f"{root_path}不是一个文件夹路径").to_str()
+        root_path = self.workspace.path_validator.validate(folder_path)
+        if not root_path.is_dir():
+            return ToolErrorResponse(self.__class__.__name__, f"{root_path}不是一个文件夹路径").to_str()
 
-            return [
-                f"{'[Folder]' if item.is_dir() else '[File]'} {item.relative_to(self.workspace.root_path)}"
-                for item in islice(root_path.glob(pattern), max_ret)
-            ]
-        except PathNotFoundError as err1:
-            return ToolErrorResponse(self.__class__.__name__, err1).to_str()
-        except WorkspaceBoundaryError as err2:
-            return ToolErrorResponse(self.__class__.__name__, err2).to_str()
-        except PermissionError as err3:
-            return ToolErrorResponse(self.__class__.__name__, err3).to_str()
-        except Exception as err:
-            return ToolErrorResponse(self.__class__.__name__, err).to_str()
+        return [
+            f"{'[Folder]' if item.is_dir() else '[File]'} {item.relative_to(self.workspace.root_path)}"
+            for item in islice(root_path.glob(pattern), max_ret)
+        ]
