+++ src/workspace/tools/read_tool.py
@@ -12,6 +12,7 @@ class ReadTool(BaseTool):
         self.func = self.read
         self.params = BaseTool.extract_params(self.read)
 
+    @BaseTool.handle_tool_exceptions
     def read(self, file_path: str, max_lines: int = 0, encoding: str = "utf-8") -> str:
         """
         读取文件内容,可限制最大行数,返回文件内容字符串(带行号)
@@ -22,38 +23,27 @@ class ReadTool(BaseTool):
         max_lines: 最大行数(0表示不限制)
         encoding: 编码
         """
-        try:
-            path: Path = self.workspace.path_validator.validate(file_path)
+        path: Path = self.workspace.path_validator.validate(file_path)
 
-            if not path.is_file():
-                return ToolErrorResponse(
-                    self.__class__.__name__, ValueError(f"读取文件{path}时未读取到完整文件")
-                ).to_str()
+        if not path.is_file():
+            return ToolErrorResponse(self.__class__.__name__, ValueError(f"读取文件{path}时未读取到完整文件")).to_str()
 
-            with open(path, encoding=encoding) as f:
-                lines = f.readlines()
+        with open(path, encoding=encoding) as f:
+            lines = f.readlines()
 
-            total_lines = len(lines)
+        total_lines = len(lines)
 
-            if max_lines > 0:
-                lines = lines[:max_lines]
+        if max_lines > 0:
+            lines = lines[:max_lines]
 
-            result_lines = []
-            for i, line in enumerate(lines, 1):
-                content = line.rstrip("\n\r")
-                result_lines.append(f"{i:6d} | {content}")
+        result_lines = []
+        for i, line in enumerate(lines, 1):
+            content = line.rstrip("\n\r")
+            result_lines.append(f"{i:6d} | {content}")
 
-            header = f"\n[文件: {path}]\n[行 1-{len(lines)} / 共 {total_lines} 行]\n"
-            separator = "-" * 80 + "\n"
+        header = f"\n[文件: {path}]\n[行 1-{len(lines)} / 共 {total_lines} 行]\n"
+        separator = "-" * 80 + "\n"
 
-            self._record_read_meta(path)
+        self._record_read_meta(path)
 
-            return header + separator + "\n".join(result_lines)
-        except PathNotFoundError as err1:
-            return ToolErrorResponse(self.__class__.__name__, err1).to_str()
-        except WorkspaceBoundaryError as err2:
-            return ToolErrorResponse(self.__class__.__name__, err2).to_str()
-        except PermissionError as err3:
-            return ToolErrorResponse(self.__class__.__name__, err3).to_str()
-        except Exception as err:
-            return ToolErrorResponse(self.__class__.__name__, err).to_str()
+        return header + separator + "\n".join(result_lines)
